Extract TreeViewBuilder
